move getTextWidth to visUtils
