+++ superset/assets/javascripts/modules/visUtils.js
@@ -1,0 +1,7 @@
+export function getTextWidth(text, fontDetails = '12px Roboto') {
+  const canvas = document.createElement('canvas');
+  const context = canvas.getContext('2d');
+  context.font = fontDetails;
+  const metrics = context.measureText(text);
+  return metrics.width;
+}
